requirements.txt
pydantic-ai-slim[openai]==0.6.0
openai>=1.92,<1.99
pydantic>=2.6,<3